Remove flow, task, and project stores, replace with monolith data store
